api/dev-requirements.txt
-r requirements.txt
black==23.3.0  # https://github.com/psf/black
factory_boy==3.2.1  # https://github.com/FactoryBoy/factory_boy/commit/88458348b01c347e31cf1524a71482923d406a4f
flake8
freezegun
ipdb
ipython
isort  # https://github.com/PyCQA/isort
mongomock_motor~=0.0.29
mypy==1.9.*
opentelemetry-distro==0.43b0
opentelemetry-exporter-otlp==1.22.0
pre-commit==2.21.0  # https://github.com/pre-commit/pre-commit/releases/tag/v2.21.0
pytest  # https://github.com/pytest-dev/pytest
pytest-asyncio
pytest-cov
pytest-mock
